refactor: made audit event as feature flag and refactored tracing-kafka
