+++ backend/common_utils/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2021"
 [dependencies]
 common_enums = { path = "../common_enums" }
 anyhow = "1.0"
-rdkafka = "0.36.2"
-tracing-kafka = { path = "../tracing-kafka" }
+rdkafka = { version = "0.36", optional = true }
+tracing-kafka = { path = "../tracing-kafka", optional = true }
 tracing = "0.1.40"
 once_cell = "1.19"
 chrono = "0.4.31"
@@ -41,6 +41,7 @@ semver = { version = "1.0.26", features = ["serde"] }
 # Optional features
 [features]
 default = []
+kafka = ["dep:rdkafka", "dep:tracing-kafka"]
 async_ext = ["dep:async-trait", "dep:futures"]
 
 [dependencies.async-trait]
