+++ backend/common_utils/src/lib.rs
@@ -13,7 +13,21 @@ pub mod request;
 pub mod types;
 // Re-export commonly used items
 pub use errors::{CustomResult, ParsingError, ValidationError};
+
+#[cfg(feature = "kafka")]
 pub use event_publisher::{emit_event_with_config, init_event_publisher};
+#[cfg(not(feature = "kafka"))]
+pub fn init_event_publisher(_config: &events::EventConfig) -> CustomResult<(), ()> {
+    Ok(())
+}
+#[cfg(not(feature = "kafka"))]
+pub async fn emit_event_with_config(
+    _event: events::Event,
+    _config: &events::EventConfig,
+) -> CustomResult<bool, ()> {
+    Ok(false)
+}
+
 pub use global_id::{CellId, GlobalPaymentId};
 pub use id_type::{CustomerId, MerchantId};
 pub use pii::{Email, SecretSerdeValue};
@@ -26,6 +40,7 @@ pub mod events;
 pub mod global_id;
 
 pub mod consts;
+#[cfg(feature = "kafka")]
 pub mod event_publisher;
 
 fn generate_ref_id_with_default_length<const MAX_LENGTH: u8, const MIN_LENGTH: u8>(
