+++ backend/grpc-server/Cargo.toml
@@ -76,7 +76,7 @@ reqwest = { version = "0.11", features = ["json"] }
 
 
 [features]
-default = ["kafka"]
+default = []
 kafka = ["tracing-kafka"]
 
 [lints]
