+++ backend/tracing-kafka/Cargo.toml
@@ -11,10 +11,17 @@ rdkafka = "0.36"
 serde_json = "1.0"
 tokio = "1.0"
 thiserror = "1.0"
-prometheus = "0.13"
-once_cell = "1.19"
 log_utils = { git = "https://github.com/juspay/framework-libs-rs", rev = "243562041252fe5897ce888d20b715ffdc3767ce", package = "log_utils", features = ["tracing"] }
+
+# Optional dependencies for metrics
+once_cell = { version = "1.19", optional = true }
+prometheus = { version = "0.13", optional = true }
 
 [lib]
 name = "tracing_kafka"
 path = "src/lib.rs"
+
+
+[features]
+default = ["kafka-metrics"]
+kafka-metrics = ["dep:once_cell", "dep:prometheus"]
